+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [0.6.4] - 2026-04-25
+
+### Added
+
+- `--ext` -- harvest files with specific extensions
+
+### Changed
+
+- `--max` -- count only harvestable/harvested files (skip “skipped” entries)
+
 ## [0.6.3] - 2026-04-24
 
 ### Added